requirements.txt
discord.py
python-dotenv
emoji
sympy
inputimeout
